ft-#11: Rename homepage folder
